+++ network/personal-chat/src/main/java/com/seugi/network/personalchat/request/PersonalChatCreateRequest.kt
@@ -3,6 +3,6 @@ package com.seugi.network.personalchat.request
 data class PersonalChatCreateRequest(
     val workspaceId: String,
     val roomName: String,
-    val joinUsers: List<Int>,
+    val joinUsers: List<Long>,
     val chatRoomImg: String,
 )
